rename Contents table to DeviceListTabTable to accurately reflect what it is
